+++ Pod/Core/ActorKit.h
@@ -7,10 +7,8 @@
 //
 
 #import "TBActor.h"
-#import "TBActorFuture.h"
 #import "TBActorProxySync.h"
 #import "TBActorProxyAsync.h"
-#import "TBActorProxyFuture.h"
 #import "TBActorPool.h"
 #import "TBActorRegistry.h"
 #import "NSException+ActorKit.h"