+++ Pod/Futures/Futures.h
@@ -1,0 +1,12 @@
+//
+//  Futures.h
+//  ActorKitFutures
+//
+//  Created by Julian Krumow on 13.08.15.
+//  Copyright (c) 2015 Julian Krumow. All rights reserved.
+//
+
+#import "TBActor.h"
+#import "TBActorProxy.h"
+#import "TBActorFuture.h"
+#import "TBActorProxyFuture.h"